:recycle: clean code
