+++ src/main/java/com/tale/controller/admin/ArticleController.java
@@ -4,39 +4,29 @@ import com.blade.exception.ValidatorException;
 import com.blade.ioc.annotation.Inject;
 import com.blade.kit.DateKit;
 import com.blade.kit.StringKit;
-import com.blade.mvc.annotation.*;
+import com.blade.mvc.annotation.Param;
+import com.blade.mvc.annotation.Path;
+import com.blade.mvc.annotation.PostRoute;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.extension.Commons;
-import com.tale.bootstrap.TaleConst;
 import com.tale.model.dto.LogActions;
 import com.tale.model.dto.Types;
 import com.tale.model.entity.Contents;
 import com.tale.model.entity.Logs;
-import com.tale.model.entity.Metas;
 import com.tale.model.entity.Users;
 import com.tale.service.ContentsService;
 import com.tale.service.MetasService;
 import com.tale.service.SiteService;
-import com.tale.validators.CommentValidator;
-import io.github.biezhi.anima.enums.OrderBy;
-import io.github.biezhi.anima.page.Page;
+import com.tale.validators.CommonValidator;
 import lombok.extern.slf4j.Slf4j;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.List;
-import java.util.Optional;
-
-import static io.github.biezhi.anima.Anima.select;
 
 /**
  * 文章管理控制器
  * Created by biezhi on 2017/2/21.
  */
 @Slf4j
-@Path("admin/article")
+@Path(value = "admin/article", restful = true)
 public class ArticleController extends BaseController {
 
     @Inject
@@ -49,68 +39,12 @@ public class ArticleController extends BaseController {
     private SiteService siteService;
 
     /**
-     * 文章管理首页
-     *
-     * @param page
-     * @param limit
-     * @param request
-     * @return
-     */
-    @GetRoute(value = "")
-    public String index(@Param(defaultValue = "1") Integer page, @Param(defaultValue = "15") Integer limit,
-                        Request request) {
-
-        Page<Contents> articles = select().from(Contents.class).where(Contents::getType, Types.ARTICLE).order(Contents::getCreated, OrderBy.DESC).page(page, limit);
-        request.attribute("articles", articles);
-        return "admin/article_list";
-    }
-
-    /**
-     * 文章发布页面
-     *
-     * @param request
-     * @return
-     */
-    @GetRoute(value = "publish")
-    public String newArticle(Request request) {
-        List<Metas> categories = metasService.getMetas(Types.CATEGORY);
-        request.attribute("categories", categories);
-        request.attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
-        request.attribute("now", DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm").format(LocalDateTime.now()));
-        return "admin/article_edit";
-    }
-
-    /**
-     * 文章编辑页面
-     *
-     * @param cid
-     * @param request
-     * @return
-     */
-    @GetRoute(value = "/:cid")
-    public String editArticle(@PathParam String cid, Request request) {
-        Optional<Contents> contents = contentsService.getContents(cid);
-        if (!contents.isPresent()) {
-            return render_404();
-        }
-        request.attribute("contents", contents.get());
-        List<Metas> categories = metasService.getMetas(Types.CATEGORY);
-        request.attribute("categories", categories);
-        request.attribute("active", "article");
-        request.attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
-        return "admin/article_edit";
-    }
-
-    /**
      * 发布文章操作
-     *
-     * @return
      */
-    @PostRoute(value = "publish")
-    @JSON
-    public RestResponse<?> publishArticle(Contents contents) {
+    @PostRoute("publish")
+    public RestResponse<?> publish(Contents contents) {
 
-        CommentValidator.valid(contents);
+        CommonValidator.valid(contents);
 
         Users users = this.user();
         contents.setType(Types.ARTICLE);
@@ -140,17 +74,14 @@ public class ArticleController extends BaseController {
 
     /**
      * 修改文章操作
-     *
-     * @return
      */
-    @PostRoute(value = "modify")
-    @JSON
-    public RestResponse<?> modifyArticle(Contents contents, @Param String createTime) {
+    @PostRoute("modify")
+    public RestResponse<?> modify(Contents contents, @Param String createTime) {
         try {
             if (null == contents || null == contents.getCid()) {
                 return RestResponse.fail("缺少参数，请重试");
             }
-            CommentValidator.valid(contents);
+            CommonValidator.valid(contents);
 
             if (StringKit.isNotBlank(createTime)) {
                 int unixTime = DateKit.toUnix(createTime, "yyyy-MM-dd HH:mm");
@@ -172,13 +103,8 @@ public class ArticleController extends BaseController {
 
     /**
      * 删除文章操作
-     *
-     * @param cid
-     * @param request
-     * @return
      */
-    @Route(value = "delete")
-    @JSON
+    @PostRoute("delete")
     public RestResponse<?> delete(@Param int cid, Request request) {
         try {
             contentsService.delete(cid);
@@ -195,4 +121,5 @@ public class ArticleController extends BaseController {
         }
         return RestResponse.ok();
     }
+
 }
