+++ src/main/java/com/tale/controller/admin/AuthController.java
@@ -4,21 +4,20 @@ import com.blade.exception.ValidatorException;
 import com.blade.kit.DateKit;
 import com.blade.kit.EncryptKit;
 import com.blade.kit.StringKit;
-import com.blade.mvc.annotation.JSON;
 import com.blade.mvc.annotation.Path;
-import com.blade.mvc.annotation.Route;
-import com.blade.mvc.http.HttpMethod;
+import com.blade.mvc.annotation.PostRoute;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.http.Response;
 import com.blade.mvc.http.Session;
 import com.blade.mvc.ui.RestResponse;
-import com.tale.controller.BaseController;
 import com.tale.bootstrap.TaleConst;
+import com.tale.controller.BaseController;
 import com.tale.model.dto.LogActions;
 import com.tale.model.entity.Logs;
 import com.tale.model.entity.Users;
 import com.tale.model.param.LoginParam;
 import com.tale.utils.TaleUtils;
+import com.tale.validators.CommonValidator;
 import lombok.extern.slf4j.Slf4j;
 
 import static io.github.biezhi.anima.Anima.select;
@@ -28,27 +27,19 @@ import static io.github.biezhi.anima.Anima.select;
  * Created by biezhi on 2017/2/21.
  */
 @Slf4j
-@Path("admin")
+@Path(value = "admin", restful = true)
 public class AuthController extends BaseController {
 
-    @Route(value = "login", method = HttpMethod.GET)
-    public String login(Response response) {
-        if (null != this.user()) {
-            response.redirect("/admin/index");
-            return null;
-        }
-        return "admin/login";
-    }
-
-    @Route(value = "login", method = HttpMethod.POST)
-    @JSON
+    @PostRoute("login")
     public RestResponse<?> doLogin(LoginParam loginParam, Request request,
-                                Session session, Response response) {
+                                   Session session, Response response) {
+
+        CommonValidator.valid(loginParam);
 
         Integer error_count = cache.get("login_error_count");
         try {
             error_count = null == error_count ? 0 : error_count;
-            if (null != error_count && error_count > 3) {
+            if (error_count > 3) {
                 return RestResponse.fail("您输入密码已经错误超过3次，请10分钟后尝试");
             }
