+++ src/main/java/com/tale/controller/admin/PageController.java
@@ -2,13 +2,13 @@ package com.tale.controller.admin;
 
 import com.blade.exception.ValidatorException;
 import com.blade.ioc.annotation.Inject;
-import com.blade.mvc.annotation.*;
+import com.blade.mvc.annotation.Param;
+import com.blade.mvc.annotation.Path;
+import com.blade.mvc.annotation.Route;
 import com.blade.mvc.http.HttpMethod;
 import com.blade.mvc.http.Request;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.controller.BaseController;
-import com.tale.extension.Commons;
-import com.tale.bootstrap.TaleConst;
 import com.tale.model.dto.LogActions;
 import com.tale.model.dto.Types;
 import com.tale.model.entity.Contents;
@@ -16,14 +16,8 @@ import com.tale.model.entity.Logs;
 import com.tale.model.entity.Users;
 import com.tale.service.ContentsService;
 import com.tale.service.SiteService;
-import com.tale.validators.CommentValidator;
-import io.github.biezhi.anima.enums.OrderBy;
-import io.github.biezhi.anima.page.Page;
+import com.tale.validators.CommonValidator;
 import lombok.extern.slf4j.Slf4j;
-
-import java.util.Optional;
-
-import static io.github.biezhi.anima.Anima.select;
 
 /**
  * 页面管理
@@ -31,7 +25,7 @@ import static io.github.biezhi.anima.Anima.select;
  * Created by biezhi on 2017/2/21.
  */
 @Slf4j
-@Path("admin/page")
+@Path(value = "admin/page", restful = true)
 public class PageController extends BaseController {
 
     @Inject
@@ -40,35 +34,10 @@ public class PageController extends BaseController {
     @Inject
     private SiteService siteService;
 
-    @Route(value = "", method = HttpMethod.GET)
-    public String index(Request request) {
-        Page<Contents> contentsPage = select().from(Contents.class).where(Contents::getType, Types.PAGE).order(Contents::getCreated, OrderBy.DESC).page(1, TaleConst.MAX_POSTS);
-        request.attribute("articles", contentsPage);
-        return "admin/page_list";
-    }
-
-    @Route(value = "new", method = HttpMethod.GET)
-    public String newPage(Request request) {
-        request.attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
-        return "admin/page_edit";
-    }
-
-    @Route(value = "/:cid", method = HttpMethod.GET)
-    public String editPage(@PathParam String cid, Request request) {
-        Optional<Contents> contents = contentsService.getContents(cid);
-        if (!contents.isPresent()) {
-            return render_404();
-        }
-        request.attribute("contents", contents.get());
-        request.attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
-        return "admin/page_edit";
-    }
-
     @Route(value = "publish", method = HttpMethod.POST)
-    @JSON
     public RestResponse<?> publishPage(Contents contents) {
 
-        CommentValidator.valid(contents);
+        CommonValidator.valid(contents);
 
         Users users = this.user();
         contents.setType(Types.PAGE);
@@ -90,9 +59,8 @@ public class PageController extends BaseController {
     }
 
     @Route(value = "modify", method = HttpMethod.POST)
-    @JSON
     public RestResponse<?> modifyArticle(Contents contents) {
-        CommentValidator.valid(contents);
+        CommonValidator.valid(contents);
 
         if (null == contents || null == contents.getCid()) {
             return RestResponse.fail("缺少参数，请重试");
@@ -114,7 +82,6 @@ public class PageController extends BaseController {
     }
 
     @Route(value = "delete")
-    @JSON
     public RestResponse<?> delete(@Param int cid, Request request) {
         try {
             contentsService.delete(cid);
@@ -131,4 +98,5 @@ public class PageController extends BaseController {
         }
         return RestResponse.ok();
     }
+
 }
