+++ src/main/java/com/tale/validators/CommonValidator.java
@@ -1,6 +1,7 @@
 package com.tale.validators;
 
 import com.blade.kit.StringKit;
+import com.blade.validator.SimpleValidation;
 import com.blade.validator.Validators;
 import com.tale.model.entity.Comments;
 import com.tale.model.entity.Contents;
@@ -13,11 +14,11 @@ import static com.tale.bootstrap.TaleConst.MAX_TITLE_COUNT;
  * @author biezhi
  * @date 2018/6/3
  */
-public class CommentValidator {
+public class CommonValidator {
 
     public static void valid(Comments param) {
         Validators.notEmpty().test(param.getAuthor()).throwIfInvalid("评论作者");
-        Validators.moreThan(30).test(param.getAuthor()).throwIfInvalid("姓名");
+        Validators.lessThan(30).test(param.getAuthor()).throwIfInvalid("姓名");
 
         Validators.notEmpty().test(param.getMail()).throwIfInvalid("电子邮箱");
         Validators.isEmail().test(param.getMail()).throwIfInvalid("电子邮箱");
@@ -27,7 +28,8 @@ public class CommentValidator {
         }
 
         Validators.notEmpty().test(param.getContent()).throwIfInvalid("内容");
-        Validators.between(5, 2000).test(param.getContent()).throwIfInvalid("内容");
+        Validators.moreThan(5).test(param.getContent()).throwIfInvalid("内容");
+        Validators.lessThan(2000).test(param.getContent()).throwIfInvalid("内容");
 
         Validators.notNull().test(param.getCid()).throwIfInvalid("评论文章");
 
@@ -41,7 +43,7 @@ public class CommentValidator {
         Validators.moreThan(MAX_TEXT_COUNT).test(param.getContent()).throwIfInvalid("文章内容");
     }
 
-    public static void valid(LoginParam param){
+    public static void valid(LoginParam param) {
         Validators.notEmpty().test(param.getUsername()).throwIfInvalid("用户名");
         Validators.notEmpty().test(param.getPassword()).throwIfInvalid("密码");
     }
